requirements.txt
# AI Standards Assistant - Dependencies

# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# PDF processing
PyMuPDF==1.23.8
pdfplumber==0.10.3
pdf2image==1.16.3

# OCR
pytesseract==0.3.10
Pillow==10.1.0

# Machine Learning & Vector Search
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
torch==2.1.1

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0

# Development (optional)
pytest==7.4.3
black==23.11.0
flake8==6.1.0
